FlaskApps/datastore-app/requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.0
Flask_Bcrypt==0.7.1
flask_talisman==0.7.0